requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# C:\Users\DPS0340\Programming\cleanerbot\src\cleaner.py: 6,7
aiohttp == 3.7.4.post0

# C:\Users\DPS0340\Programming\cleanerbot\src\cleaner.py: 8,12
# C:\Users\DPS0340\Programming\cleanerbot\src\main.py: 2,3
discord.py == 1.7.3

# C:\Users\DPS0340\Programming\cleanerbot\src\cleaner.py: 9
pyquery == 1.4.3

brotlipy == 0.7.0